MAINT,BLD: Move to a common folder
This makes calls ever so slightly slower, by 0.05 ns typically because
of the library call..
