+++ client/potentials/Aluminum/meson.build
@@ -7,7 +7,6 @@ aluminum = shared_library('aluminum',
            'gagafeDblexp.f',
            'embedenergy.f',
            'sumembforce.f',
-           '../../PotHelpers.cc',
            dependencies : _deps,
            cpp_args : _args,
            fortran_args : _fargs,
