+++ client/potentials/LJ/meson.build
@@ -1,5 +1,5 @@
 lennard_jones = shared_library('lennard_jones',
-           'LJ.cpp', '../../PotHelpers.cc',
+           'LJ.cpp',
            dependencies : _deps,
            cpp_args : _args,
            link_with : _linkto,
