+++ client/potentials/Morse/meson.build
@@ -1,6 +1,5 @@
 morse = shared_library('morse',
            'Morse.cpp',
-           '../../PotHelpers.cc',
            dependencies : _deps,
            cpp_args : _args,
            link_with : _linkto,
