+++ client/potentials/PotHelpers.cc
